fix: a bunch of new clippy lints (#1835)
* fix: a bunch of new clippy lints

* Update fig_desktop/src/webview/window.rs

* restore original

* fix ubuntu

* more fixes

* more lints

+++ rust-lib/alacritty_terminal/src/index.rs
@@ -59,7 +59,7 @@ pub struct Rect {
 
 impl Rect {
     pub const fn size(&self) -> usize {
-        (self.end.column.0 - self.start.column.0) as usize * (self.end.line.0 - self.start.line.0) as usize
+        (self.end.column.0 - self.start.column.0) * (self.end.line.0 - self.start.line.0) as usize
     }
 }
